+++ frontend/src/layouts/AuthLayout.tsx
@@ -1,6 +1,6 @@
 import { ColorModeButton } from '@/components/buttons/ColorModeButton';
 import { Link, useNavigate } from 'react-router-dom';
-import { MobileNav } from '@/components/navigation/MobileNav';
+import { MobileNav } from '@/components/nav/MobileNav';
 import { TOKEN_NAME } from '@/config';
 
 export default function AuthLayout({ children }: { children: React.ReactNode }) {